# docs/auth_service_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Saturn MouseHunter Auth Service",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/admin/roles/": {
      "post": {
        "tags": [
          "角色管理"
        ],
        "summary": "Create Role",
        "description": "创建角色",
        "operationId": "create_role_api_v1_admin_roles__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "角色管理"
        ],
        "summary": "List Roles",
        "description": "获取角色列表",
        "operationId": "list_roles_api_v1_admin_roles__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "role_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Role Name"
            }
          },
          {
            "name": "role_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Role Code"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RoleScope"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scope"
            }
          },
          {
            "name": "is_system_role",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is System Role"
            }
          },
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Active"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoleOut"
                  },
                  "title": "Response List Roles Api V1 Admin Roles  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/roles/count": {
      "get": {
        "tags": [
          "角色管理"
        ],
        "summary": "Count Roles",
        "description": "获取角色总数",
        "operationId": "count_roles_api_v1_admin_roles_count_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "role_name",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Role Name"
            }
          },
          {
            "name": "role_code",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Role Code"
            }
          },
          {
            "name": "scope",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RoleScope"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scope"
            }
          },
          {
            "name": "is_system_role",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is System Role"
            }
          },
          {
            "name": "is_active",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Is Active"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 100,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "minimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Count Roles Api V1 Admin Roles Count Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/roles/system": {
      "get": {
        "tags": [
          "角色管理"
        ],
        "summary": "List System Roles",
        "description": "获取系统角色列表",
        "operationId": "list_system_roles_api_v1_admin_roles_system_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RoleOut"
                  },
                  "type": "array",
                  "title": "Response List System Roles Api V1 Admin Roles System Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/roles/{role_id}": {
      "get": {
        "tags": [
          "角色管理"
        ],
        "summary": "Get Role",
        "description": "获取角色详情",
        "operationId": "get_role_api_v1_admin_roles__role_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Role Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "角色管理"
        ],
        "summary": "Update Role",
        "description": "更新角色",
        "operationId": "update_role_api_v1_admin_roles__role_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Role Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "角色管理"
        ],
        "summary": "Delete Role",
        "description": "删除角色",
        "operationId": "delete_role_api_v1_admin_roles__role_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Role Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/roles/{role_id}/permissions": {
      "get": {
        "tags": [
          "角色管理"
        ],
        "summary": "Get Role With Permissions",
        "description": "获取角色及其权限",
        "operationId": "get_role_with_permissions_api_v1_admin_roles__role_id__permissions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Role Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoleWithPermissions"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PermissionOut": {
        "properties": {
          "permission_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Permission Name",
            "description": "权限名称"
          },
          "permission_code": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Permission Code",
            "description": "权限编码"
          },
          "resource": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Resource",
            "description": "资源名称"
          },
          "action": {
            "type": "string",
            "maxLength": 50,
            "minLength": 2,
            "title": "Action",
            "description": "操作类型"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "权限描述"
          },
          "is_system_permission": {
            "type": "boolean",
            "title": "Is System Permission",
            "description": "是否系统权限",
            "default": false
          },
          "id": {
            "type": "string",
            "title": "Id",
            "description": "权限ID"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "创建时间"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "更新时间"
          }
        },
        "type": "object",
        "required": [
          "permission_name",
          "permission_code",
          "resource",
          "action",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "PermissionOut",
        "description": "权限输出模型"
      },
...
      "RoleIn": {
        "properties": {
          "role_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Role Name",
            "description": "角色名称"
          },
          "role_code": {
            "type": "string",
            "maxLength": 50,
            "minLength": 2,
            "title": "Role Code",
            "description": "角色编码"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "角色描述"
          },
          "scope": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RoleScope"
              }
            ],
            "description": "角色范围",
            "default": "GLOBAL"
          },
          "is_system_role": {
            "type": "boolean",
            "title": "Is System Role",
            "description": "是否系统角色",
            "default": false
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "是否激活",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "role_name",
          "role_code"
        ],
        "title": "RoleIn",
        "description": "角色创建模型"
      },
      "RoleOut": {
        "properties": {
          "role_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Role Name",
            "description": "角色名称"
          },
          "role_code": {
            "type": "string",
            "maxLength": 50,
            "minLength": 2,
            "title": "Role Code",
            "description": "角色编码"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "角色描述"
          },
          "scope": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RoleScope"
              }
            ],
            "description": "角色范围",
            "default": "GLOBAL"
          },
          "is_system_role": {
            "type": "boolean",
            "title": "Is System Role",
            "description": "是否系统角色",
            "default": false
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "是否激活",
            "default": true
          },
          "id": {
            "type": "string",
            "title": "Id",
            "description": "角色ID"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "创建时间"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "更新时间"
          }
        },
        "type": "object",
        "required": [
          "role_name",
          "role_code",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "RoleOut",
        "description": "角色输出模型"
      },
      "RoleScope": {
        "type": "string",
        "enum": [
          "GLOBAL",
          "TENANT",
          "SYSTEM"
        ],
        "title": "RoleScope",
        "description": "角色范围"
      },
      "RoleUpdate": {
        "properties": {
          "role_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 2
              },
              {
                "type": "null"
              }
            ],
            "title": "Role Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "scope": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RoleScope"
              },
              {
                "type": "null"
              }
            ]
          },
          "is_active": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Active"
          }
        },
        "type": "object",
        "title": "RoleUpdate",
        "description": "角色更新模型"
      },
      "RoleWithPermissions": {
        "properties": {
          "role_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Role Name",
            "description": "角色名称"
          },
          "role_code": {
            "type": "string",
            "maxLength": 50,
            "minLength": 2,
            "title": "Role Code",
            "description": "角色编码"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "角色描述"
          },
          "scope": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RoleScope"
              }
            ],
            "description": "角色范围",
            "default": "GLOBAL"
          },
          "is_system_role": {
            "type": "boolean",
            "title": "Is System Role",
            "description": "是否系统角色",
            "default": false
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active",
            "description": "是否激活",
            "default": true
          },
          "id": {
            "type": "string",
            "title": "Id",
            "description": "角色ID"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At",
            "description": "创建时间"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At",
            "description": "更新时间"
          },
          "permissions": {
            "items": {
              "$ref": "#/components/schemas/PermissionOut"
            },
            "type": "array",
            "title": "Permissions",
            "description": "角色权限列表"
          }
        },
        "type": "object",
        "required": [
          "role_name",
          "role_code",
          "id",
          "created_at",
          "updated_at"
        ],
        "title": "RoleWithPermissions",
        "description": "包含权限的角色模型"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}